Lexer: implemented parsing arrays
Now after parsing a word, posInLine points to the last symbol of the
word, so that the lexer can continue processing the input stream from
the next symbol.

Changed the lexer so that it allows an ID to have digits in the middle
or end of the word.

Any non-digit symbol after a number is parsed separately after the
number

Fixed an error: posInLine now also counts spaces